requirements.txt
pydantic-settings
google-genai
playwright